requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pyopenssl
json2html
python-ssl-checker
pydantic_settings
